+++ src/contextTypes.js
@@ -1,4 +1,4 @@
-import {object, number, instanceOf} from 'prop-types'
+import {object, instanceOf} from 'prop-types'
 import jss, {SheetsRegistry} from './jss'
 import * as ns from './ns'
 
@@ -6,5 +6,5 @@ export default {
   [ns.jss]: instanceOf(jss.constructor),
   [ns.sheetOptions]: object,
   [ns.sheetsRegistry]: instanceOf(SheetsRegistry),
-  [ns.providerId]: number
+  [ns.managers]: object
 }