+++ changelog.md
@@ -1,8 +1,11 @@
 ## 8.0.0 / 2017-09-30
 
 - Upgrade to JSS 9.
-- Pass component name as a classNamePrefix.
-- Added classNamePrefix prop to JssProvider.
+- Fixed manager leaking to next request on SSR (#133).
+- Pass component name as a classNamePrefix (#133).
+- Added classNamePrefix prop to JssProvider for better multi-tree support (#133).
+- Allowed nesting of JssProvider (#157).
+- Allow JssProvider rerender when used deep in the tree (#128).
 
 ## 7.2.0 / 2017-09-23